+++ src/App.js
@@ -1,14 +1,12 @@
 import React from "react";
 import { HashRouter, Switch, Route, Redirect } from "react-router-dom";
 import { Navigation } from "./common/Navigation";
-
 import Descritpion from "./features/Description";
 import MovieDetails from "./features/movieDetails";
 import MovieList from "./features/movieList";
 import { People } from "./features/people/PeoplePage";
 
 export const App = () => {
-
 
   return (
     <>
@@ -41,6 +39,5 @@ export const App = () => {
     </>
   );
 };
-
 
 export default App;
