+++ src/GlobalStyle.js
@@ -3,15 +3,15 @@ import { createGlobalStyle } from "styled-components";
 export const GlobalStyle = createGlobalStyle`
     html {
         box-sizing: border-box;
-    }
+    };
 
     *,::after, ::before {
         box-sizing: inherit;
-    }
+    };
 
     body {
         margin: 0 auto;
-        background: ${({theme}) => theme.color.whisper};
+        background: ${({ theme }) => theme.color.whisper};
         font-family: "Poppins", sans-serif;
-    }
+    };
 `
