+++ src/common/NoResultPage/index.js
@@ -3,7 +3,7 @@ import { useParams } from "react-router-dom";
 import { selectMovieTotalPages } from "../../features/movieList/movieListSlice";
 import { useQueryParameter } from "../../queryParameters";
 import searchQueryParamName from "../../searchQueryParamName";
-import { ReactComponent as NoResultIcon } from "./NoResultIcon/picture.svg";
+import { ReactComponent as NoResultIcon } from "../../images/picture.svg";
 import { Container, StyledHeader, Wrapper } from "./styled";
 
 const NoResultPage = () => {
@@ -14,7 +14,7 @@ const NoResultPage = () => {
     return (
         <>
             <Wrapper>
-                <StyledHeader>{(page > totalPages && totalPages != 0) ? `Sorry, there are no more result for "${query}". Please check previous pages.` : `Sorry, there are no result for "${query}"`}</StyledHeader>
+                <StyledHeader>{(Number(page) > totalPages && totalPages !== 0) ? `Sorry, there are no more result for "${query}". Please check previous pages.` : `Sorry, there are no result for "${query}"`}</StyledHeader>
                 <Container>
                     <NoResultIcon />
                 </Container>
