+++ src/features/movieList/pagination/PaginationNumberPage/index.js
@@ -1,12 +1,12 @@
 import React from "react";
 import { useSelector } from "react-redux";
-import { useParams } from "react-router-dom";
 import { theme } from "../../../../theme";
 import { selectMovieListState } from "../../movieListSlice";
 import { PageNumberBox, Number } from "./styled";
 
 const PaginationPageNumber = () => {
     const { page, lastPage, totalPages } = useSelector(selectMovieListState);
+
     return (
         <PageNumberBox theme={theme}>
             Page
@@ -15,5 +15,5 @@ const PaginationPageNumber = () => {
             <Number>{totalPages < lastPage ? totalPages : lastPage}</Number>
         </PageNumberBox>
     )
-}
+};
 export default PaginationPageNumber
