+++ src/getApi.js
@@ -94,3 +94,40 @@ export const searchPeople = async (query, page) => {
     };
 };
 
+export const getDetails = async (id) => {
+    const response = await fetch(
+        `https://api.themoviedb.org/3/person/${id}?api_key=9515ffc857c67f1558538dad140abb29&language=en-US`
+    );
+
+    if (!response.ok) {
+        throw new Error(response.statusText);
+    };
+
+    return await response.json();
+};
+
+export const getProjects = async (id) => {
+
+    const response = await fetch(`https://api.themoviedb.org/3/person/${id}/movie_credits?api_key=9515ffc857c67f1558538dad140abb29&language=en-US`)
+    const data = await response.json()
+
+    if (!response.ok) {
+        new Error(response.statusText)
+    }
+
+    return await data;
+};
+
+export const getPeopleList = async (peoplePage) => {
+    const response = await fetch(
+        `https://api.themoviedb.org/3/person/popular?api_key=9515ffc857c67f1558538dad140abb29&language=en-US&page=${peoplePage}`
+    );
+
+    const data = await response.json();
+
+    if (!response.ok) {
+        new Error(response.statusText);
+    }
+
+    return await data;
+};
