+++ src/queryParameters.js
@@ -1,4 +1,4 @@
-import { useHistory, useLocation, useParams } from "react-router-dom";
+import { useHistory, useLocation } from "react-router-dom";
 
 export const useQueryParameter = (key) => {
     const location = useLocation();
