+++ src/theme.js
@@ -1,4 +1,3 @@
-
 export const theme = {
     color: {
         white: "#FFFFFF",
